+++ lib/oanda_api_v20.rb
@@ -1,3 +1,7 @@
+require 'byebug'
+
+
+
 require 'httparty'
 require 'persistent_httparty'
 require 'http/exceptions'